requirements.txt
streamlit==1.32.0
pandas==2.1.3
pyodbc==5.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
pyyaml==6.0.1